Atomically delete object queue only when empty
Once the threads object queue is empty, other threads may explicitly
merge reference counts immediately. It's no longer safe for the original
thread to do RC operations because of races with accesses to ob_ref_local.

This ensures that the queue is empty when it's removed from the hashtable
mapping thread-id to queue. There are still other RC operations that can
happen during thread destruction, so the call to _Py_queue_destroy will
probably need to be moved.

See #50