Don't use core APIs in parking_lot.h
This changes the parking lot API to use PyThreadState instead of the
private PyThreadStateOS struct in function arguments. The "token" field
used for fair locking is moved to PyThreadState as "handoff_elem". It
will be used for handing off queue elements in a future commit.

_PySemaphore_Wait now returns the same error/successs codes as the
rest of the parking lot API, including PY_PARK_INTR for when the
condition variable spuriously wakes up (often due to a signal).

WIP: rework parking