Remove PySimpleQueue
The _queue.SimpleQueue class is always available now that threads are
always available. I'm too lazy to try to make the Python implementation
thread-safe.

SimpleQueue is used in setup.py so we now need to build _queue earlier.

The C implementation queue.SimpleQueue is always available and
I'm too lazy to make the Python implementation thread-safe.