Remove unused Python portions of functools
Some functions in functools exist in Python and in C. The C
implementations are used by default, although test_functools contains
test cases specifically for the Python implementation as well.

This deletes the otherwise unused Python implementations of in
preparation for making lru_cache thread-safe without the GIL. This
assumes that the C implementation is always available.