functools: make lru_cache thread-safe
This adds a recursive mutex to each LRU cache. The mutex is held
while the cache is updated, but released when calling the annotated
function. Note that this still potentially can lead to deadlocks that
did not exist with the GIL, but I think only in rather esoteric cases.

I removed the TestCase.subTest call in test_lru_cache_threaded3 because
it's not thread-safe without the GIL. I think it may also be incorrect
even with the GIL.