[bpo-46205] runtest_mp: exit if no workers are alive
I think there is a race condition in runtest_mp where if a worker
has already pushed its final output to the queue, but is still
"alive", then the main thread waits forever on the the queue.
This might fix that issue, but there's still a delay of up
to ~30 secs.

See https://bugs.python.org/issue46205