code: only use co_zombieframe when GIL is on
The co_zombieframe modifications introduce an obstacle to scalability,
due to false sharing with nearby fields.

In ceval, rearrange the GIL check and the opcache cache. This is
probably not necessary with the above changes, but I'm keeping it
because it seemed to mitigate some of the problems with false sharing
by reducing reads to nearby fields.