qsbr: remove stop-the-world call
This replaces the array of qbsr threads with a stack, which
allows us to get rid of the stop-the-world call. There was
an initialization order bug: registering a new qsbr thread
might stop-the-world, but stopping the world requires the
thread to already be attached!

This will probably make the qsbr scan even slower, but we
can improve that in the future.

Fixes #97