vm: Implement finally blocks
This adds the CALL_FINALLY opcode, which stores the address of the
next_instr in opA register and jumps to the label at opD. The
END_FINALLY instruction jumps to the address stored in the opA
register (if the value is positive), continues to the next instruction
(if the value is zero), or re-raises the exception in the opA+1
register (if the value is -1).

This is a slightly different strategy than CPython, which duplicates
the code for the finally block for each control flow statement that
passes through it (return, break, continue).