vm: avoid stack usage in RETURN_VALUE
This changes the DECREF loop in a  few ways. First, it uses an integer
instead of a "top" pointer. This fixes a bug where the "top" pointer
would be invalid if regs gets resized during a destructor call.

Second, the DECREF loop doesn't save next_instr to ts. Instead it's
next_instr is saved once before the loop. This allows the compiler
to re-use the next_instr register for "n", since next_instr isn't
used again (it's loaded from the frame link or the function exits).

This seems to speed-up fib (by ~7%) and fixes a bug.