vm: handle clean-up for FUNC_HEADER errors sepcially
The calling function typically cleans up arguments to failed function
calls, but calls from C into Python don't do this. This can leave
garbage in registers if the failed call passes too many arguments. This
adds a special function to clean-up failed calls in the caller based on
the value of the accumulator. For simplicity, we use this for all failed
calls.