Fix early exits from with/async with
When unwinding a with or async with block we inserted
END_WITH/END_ASYNC_WITH opcodes at the unwind point. The problem is that
the unwinding code is still within the exception handler -- if the
__exit__ function raises an exception, it will get called again in the
handler.

This uses CALL_FINALLY to jump to the 'finally' block containing
END_ASYNC_WITH. END_WITH is made idempotent on errors.

Adds a test to test_with.py