pystate: check "gc_thread_countdown" for stop condition
The code previously checked "runtime->stop_the_world", but this could
lead to deadlocks on exit. The EVAL_PLEASE_STOP can remain on a thread
even after the world is resumed, if another thread is responsible for
switching it to the GC state. If that thread then stops-the-world again
before cleaing EVAL_PLEASE_STOP it may deadlock when running GC
finalizers and noticing the signal.

This instead checks "gc_thread_countdown", which unlike "stop_the_world",
is guranteed to be zero by the time the thread finishes the
stop-the-world process.

The deadlock was seen in test_3_join_in_forked_from_thread when the GIL
was disabled by default.