Disable the GIL by default.
Changes "test_cancel_futures_wait_false" to pause a tiny bit before
calling "shutdown" with cancel_futures=True. Otherwise, the work item
may be cancelled before it is started. In practice, this only happens
when the GIL is disabled. With the GIL on, the thread grabs the work
item before t.shutdown() is called due to the 5 ms GIL switch interval.