vm: add jump side table to support large forward jumps
Forward jumps pose a challenge for the bytecode compiler because the
size of instruction isn't know when it's emitted. This adds a side table
in the code object containing jump addresses and targets for forward
jumps that don't fit in 16 bit offsets. Use of the side table is
indicated by a jump offset of "0".

This is a similar to the approach used by WebKit's low-level
interpreter.

Fixes #5