Use opcode_targets[0] as eval_breaker
There is no valid 0 opcode, so it's convenient to merge the first entry
of opcode_targets with the thread's eval_breaker. In particular, this
makes it easiert to convince the compiler to generate reasonable code
when checking the eval breaker. Previously, this sometimes required
BREAK_LIVE_RANGE to avoid the compiler (uncessarily) stashing the
address of eval_breaker in a non-volatile register (across calls) when
it can easily be addressed by an offset from opcode_targets.