Use types.CodeTypes instead of CodeType (temporarily)
Cython generates "fake" code objects for functions, using the PyCode_New
and PyCode_NewEmpty APIs. Those still create the "old" code objects that
are otherwise unused. That's not generally an issue, except that the
inspect module checks the type of the code object to create signatures.
This is a temporary fix to get pydantic working when built with Cython.

The better fix will be to get rid of the old code objects and make
PyCode_New and PyCode_NewEmpty return instances of the new code objects.

See #21