atomic: add bulitin (GCC) atomics
Previously, we used the frozen pyconfig.h feature flags to determine
which atomic implementation to use. This did not work for the SciPy
Windows build because it builds with GCC against a Windows Python, which
effectively assumed MSVC.

This moves the feature detection later when pyatomic.h is included and
also adds an implementation for GCC builtin atomics (also implemented by
Clang) that don't require compiling with "--std=c11".