gc: swap order of PyGC_Head and managed dict/weakref
This is a temporary change to make rebasing easier. The tracing garbage
collector needs to identify live vs. dead blocks. This was combined with
the checking of tracked objects: tracked objects always have the low bit
of PyGC_Head->_gc_prev set to 1. Unallocated mimalloc blocks always have
a pointer (or NULL) in the first word; the LSB is never 1.

Ensuring that PyGC_Head comes at the beginning of the allocation makes
this easier.