obmalloc: process queue of to-be-freed objects before thread exit
When a thread exits, it merges it's local to-be-freed queue of objects
into the interpreter's queue. If PyThreadStates are created and
destroyed frequently without doing much work or triggering a GC, this
can lead to the interpreter's queue growing large, effectively leaking
memory. This changes the thread exit code to process the local and
interpreter queue before exiting, to ensure that the interpreter queue
is processed frequently enough.