multiprocessing: remove use of CriticalLock
The use of CriticalLock could lead to deadlocks. Acquiring a
CriticalLock prevents the thread from pausing for garbage collection.
However, it might still block while attempting to acquire other locks.

The deadlock looks like:

 Thread 1:
   - holds _PyRuntime.mutex
   - blocked waiting for other threads to pause for GC

 Thread 2:
   - holds CriticalLock (won't switch to GC state)
   - blocked trying to acquire _PyRuntime.mutex

It might be tempting to try to prevent threads from holding
_PyRuntime.mutex (or other locks) across GCs, but that's not possible
with the current lock design because:

 - Threads can be switched to GC state while detached and waiting to
   acquire a lock
 - Locks can be handed off to a blocked thread (even if it's now in GC
   state).