lock: fix bug when locking PyDictOrValues (initialize 'handoff')
When locking a PyDictOrValues object, a thread may incorrectly think
it's acquired the lock due to an uninitialized handoff variable. This
was not an issue for other locking code paths because in those cases the
unlocking thread would always set handoff (i.e., it would always be
initialized before it was read).

Change the locking code to always intialize handoff to zero, even in
cases where that may not be necessary.